+++ src/api/upload/types/upload.ts
@@ -1,0 +1,5 @@
+export interface IFormInfo {
+   title: string;
+   body: string;
+   files: File[];
+}
