+++ src/pages/petition/post.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { API_PATH } from 'constants/api';
 import { ROUTES } from 'constants/route';
 import useImageUpload from 'hooks/useImageUpload';
-import { IFormInfo } from 'hooks/useFormUpload';
 import { useFormUpload } from 'hooks/useFormUpload';
 import Post from 'components/main/post';
+import { IFormInfo } from 'api/upload/types/upload';
 
 export default function PetitionForm() {
    const initFormInfo: IFormInfo = {
